backend/requirements.txt
# Core dependencies
Flask==2.2.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-JWT-Extended==4.4.4
SQLAlchemy==2.0.5
alembic==1.10.1
celery==5.2.7
gunicorn==20.1.0
python-dotenv==1.0.0
Werkzeug==2.2.3
requests==2.28.2
jsonschema==4.17.3
redis==4.5.1
psycopg2-binary==2.9.5

# For file handling and conversions
Pillow==9.4.0
pandas==1.5.3
openpyxl==3.1.1
PyYAML==6.0
lxml==4.9.2
beautifulsoup4==4.11.2
markdown==3.4.1
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
weasyprint==58.1
html2text==2020.1.16
python-docx==0.8.11
cairosvg==2.6.0
pygments==2.14.0
fpdf==1.7.2
pdf2docx==0.5.6
pdfkit==1.0.0

# Audio/Video conversion
ffmpeg-python==0.2.0
moviepy==1.0.3
pydub==0.25.1

# Testing
pytest==7.2.2
pytest-cov==4.0.0

# Development
black==23.1.0
isort==5.12.0
flake8==6.0.0
mypy==1.0.1 